+++ MoMoney/Models/Log.cs
@@ -7,17 +7,17 @@ public class Log
 {
     [PrimaryKey, AutoIncrement]
     public int LogId { get; set; }
-    public DateTime TimeStamp { get; set; }
+    public DateTime Timestamp { get; set; }
     public string Level { get; set; } = LogLevel.Information.ToString();
-    public string ClassName { get; set; } = "";
-    public string? ExceptionType { get; set; } = null;
     public string Message { get; set; } = "";
+    public string ClassName { get; set; } = "";
+    public string ExceptionType { get; set; } = "";
 
     public Log() { }
 
-    public Log(LogLevel level, string className, string message, string? exceptionType)
+    public Log(LogLevel level, string className, string message, string exceptionType)
     {
-        TimeStamp = DateTime.Now;
+        Timestamp = DateTime.Now;
         Level = level.ToString();
         ClassName = className;
         Message = message;