+++ MoMoney/Views/Settings/LoggingPage.xaml.cs
@@ -1,0 +1,13 @@
+using MoMoney.ViewModels.Settings;
+
+namespace MoMoney.Views.Settings;
+
+public partial class LoggingPage : ContentPage
+{
+	public LoggingPage(LoggingViewModel vm)
+	{
+		InitializeComponent();
+		BindingContext = vm;
+		Loaded += vm.Init;
+	}
+}